refactor: operators as provided methods on Observable instead of extension traits
