+++ src/ops/delay.rs
@@ -1,42 +1,20 @@
 use crate::prelude::*;
-use std::time::{Duration, Instant};
+use observable::observable_proxy_impl;
+use std::time::Duration;
 
-/// Delays the emission of items from the source Observable by a given timeout
-/// or until a given `Instant`.
-pub trait Delay {
-  fn delay(self, dur: Duration) -> DelayOp<Self>
-  where
-    Self: Sized,
-  {
-    DelayOp {
-      source: self,
-      delay: dur,
-    }
-  }
-  fn delay_at(self, at: Instant) -> DelayOp<Self>
-  where
-    Self: Sized,
-  {
-    DelayOp {
-      source: self,
-      delay: at.elapsed(),
-    }
-  }
-}
-impl<S> Delay for S {}
-
+#[derive(Clone)]
 pub struct DelayOp<S> {
-  source: S,
-  delay: Duration,
+  pub(crate) source: S,
+  pub(crate) delay: Duration,
 }
+
+observable_proxy_impl!(DelayOp, S);
 
 impl<S> SharedObservable for DelayOp<S>
 where
   S: SharedObservable + Send + Sync + 'static,
   S::Unsub: Send + Sync,
 {
-  type Item = S::Item;
-  type Err = S::Err;
   type Unsub = SharedSubscription;
   fn actual_subscribe<
     O: Observer<Self::Item, Self::Err> + Sync + Send + 'static,
